[REFACTOR] 모듈화, 반복문 제거
