+++ 1차/발전/interfaces/dinner.ts
@@ -1,0 +1,7 @@
+import { Member } from './member';
+
+export interface Dinner {
+  member: Member[];
+  shuffle(array: Member[]): Member[];
+  organize(array: Member[]): void;
+}
