+++ 1차/발전/interfaces/member.ts
@@ -1,0 +1,4 @@
+export interface Member {
+  name: string;
+  group: string;
+}
