+++ 1차/실천/array.js
