+++ 1차/실천/interface.ts
