+++ 1차/실천/object.js
