+++ 1차/실천/type.ts
