+++ 1차/실천/variable.js
